code-samples/eventing/bookstore-sample-app/ML-sentiment-analysis/requirements.txt
Flask==3.0.2
textblob==0.18.0.post0
parliament-functions==0.1.0